+++ lint-staged.config.js
@@ -1,3 +1,3 @@
-module.exports = {
+export default {
   '*.{js,jsx,ts,tsx,css,scss,json,md,yml,yaml}': ['prettier --write'],
 };